The cumulative fixes [BFB] for CLUBBe include
1. Workaround fix to make clubb codes run on mira by disabling XLF inlinging for two clubb modules (through Depends.mira dn Depends.cetus),
2. Bug fix to move the declaration of iSkw_zm =0 before the variable is declared as threadprivate (stats_variables.F90)
3. Bug fixes for non-standard usage of kind parameter on declarations which affect NAG compiler (endian.F90, parameters_model.F90).
4. Bug fix for a real declaration in clubb_intr.F90
5. set clubb_rainevap_turb = .false. as default and restore the default value for cldwat_icritc in FV mode. [NML] only when CLUBB is used

Define namelist parameters for CLUBB tuning

Defined CLUBB parameters in namelist_definition.xml to allow using user_nl_cam to specify the tuning values;
and revised the PGI compiler fix in stats_variables.F90 to have iSkw_zm defined before being declared threadprivate.

    [BFB]

+++ components/cam/src/physics/cam/clubb_intr.F90
@@ -483,7 +483,7 @@ end subroutine clubb_init_cnst
 
     real(kind=time_precision) :: dum1, dum2, dum3
     
-    real(8), dimension(nparams)  :: clubb_params    ! These adjustable CLUBB parameters (C1, C2 ...)
+    real(r8), dimension(nparams)  :: clubb_params    ! These adjustable CLUBB parameters (C1, C2 ...)
 
     logical :: clubb_history, clubb_rad_history, clubb_cloudtop_cooling, clubb_rainevap_turb, clubb_expldiff ! Stats enabled (T/F)
